requirements.txt
# external requirements
click~=8.1.7

coverage
# awscli
flake8
python-dotenv>=0.5.1
black

# --extra-index-url https://download.pytorch.org/whl/cu118
# torch==2.1.1+cu118
torch==2.1.1
torchvision
torchaudio


#For elastic search
elasticsearch==7.9.1

scipy
evaluate
peft
accelerate

bitsandbytes==0.41.3.post2

fasttext==0.9.2

datasets==2.13.1
fsspec==2023.9.2

protobuf
OpenCC
sacrebleu
jinja2~=3.1.2

tqdm~=4.65.0
pandas~=2.1.4
levenshtein~=0.23.0
setuptools

llama-cpp-python~=0.2.54
openai~=1.13.3
fugashi[unidic_lite]
unbabel-comet>=2.2.1
transformers~=4.38.1